Using a reentrant mutex in the appender class.
This fixes an obscure little bug in some rails applications where attempting
to log an error causes a call to the database. In turn, this database call
attempts to log causing a deadlock on the appender mutex. Since this call is
coming from the same thread of execution, a reentrant mutex solves the
problem.