Efficiency changes to the rolling file appender.
The rolling file appender uses a copy/truncate methodology for rolling the
file now. This eliminates the need for every write to the file to call stat in
order to determine if the file inode has changed.

flock is used to safely copy/truncate the file. Care must be taken when using
the rolling file appender in conjunction with fork - the rolling file appender
must be reopened after a fork to ensure the exclusive flock will function
correctly.