Calling super within a block is broken.
A regression was introduced into ruby 1.8.7 after p174 that breaks the ability
to call "super" within a block. The rolling file appender relies on this
ability to wrap up writes to the log file inside an "flock" block.

The fix is to alias the write method and then call the alias from within the
block.