Adding periodic flushing to the buffering module.
Periodic flushing allows messages to be flushed from the buffer and written to
the appender at a regular interval instead of waiting for a message threshold
to be reached. This is useful in the case of the email appender; there it
strikes a nice balance between flooding your inbox with an email for every
message and not seeing errors or failures because the threshold was never hit.