The method used to close the IO stream is now configurable.
For testing purposes (at least) we only want to close the write end of the
StringIO appender and leave the read end open. The solution is to make
configurable the close method used by the IO appender when it is closing the
appender destination. This way we can close the write end of the StringIO but
still read any log messages that were written.