Properly handling references to parent threads.
We do not want a reference to the parent thread to end up in the binding
context of the block we use to set the diagnostic contexts of the child
thread. Any reference in that binding will prevent the parent thread from
being properly garbage collected when the parent thread exits.

The same is true for the diagnostic contexts themselves. Hence we created
duplicates which get passed to the child thread.