Enable asynchronous writes of the log buffer
These changes enable the existing periodic flushing mechanism to be used to
asynchronously write the log buffer to the IO stream. By setting the `async`
flag to true, the flusher Thread will be used to write messages thereby freeing
up the main program thread to continue on doing other work.