Fix race condition in mtime calculation
The code as it stood before would first check to see if the `copy_file`
exists, and then the `mtime` of the `copy_file` would be requested.
There are some conditions where the `copy_file` could be deleted between
the existence check and the `mtime` request.

The change here just grabs the `mtime` and handles the ENOENT exception
if one arises.